+++ events.js
@@ -1,5 +1,0 @@
-const events = {
-  CHANGE_SERVICE_POINT: 'CHANGE_SERVICE_POINT'
-};
-
-export default events;