+++ CHANGELOG.md
@@ -3,4 +3,6 @@
 ## [1.0.0] (IN PROGRESS)
 
 * First release.
+* Add ability to close SP Modal. Part of UIU-628.
+* Added `onClose` callback to `ServicePoints`. Supports UIU-551.
 * Add ability to close SP Modal. Highlight currently selected service point. Part of UISP-1.